Create models folder and move appropriate files there
