+++ app/models/energy_system/xmlresource.py
