+++ tests/models/energy_system/tests.py
@@ -1,0 +1,5 @@
+'''
+TODO
+This folder mostly contains code that was already here, without tests.
+Should spent some time going through it all and write tests for it!
+'''
