+++ tests/models/test_esdl_to_scenario_converter.py
@@ -1,0 +1,3 @@
+''' Tests for the converter '''
+
+# from app.models.esdl_to_scenario_converter import EsdlToScenarioConverter
